+++ src/view/CarRaceView.js
